Add support for hardware version 1.8
This has currently no side effects. Maybe we add API for the new
POWER_GOOD pin, maybe we change the voltage divider in the future
and we need the version check for the voltage calculation.

But it is also possible that we don't do anything with the new pins,
in this case there is no need to actually release this.