Add support for Raspberry Pi 5. Auto-generate EEPROM patch.
RPi 5 requires 'compatible = "brcm,bcm2712";' in the dts files. Also
on a Pi 5 using the CS pin as GPIO is not allowed anymore. To fix this,
disable the spi0's CS pin.

This commit also adds a script that auto-generates the EEPROM patch to
switch between the DS1307 and PCF8523 variants of the device tree
overlay.